requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
redis==6.2.0
sqlparse==0.5.3
